slightly refactored, less pub
